delete old config files
