--- .babelrc
@@ -1,57 +1,0 @@
-{
-  "env": {
-    "development": {
-      "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "loose": true,
-            "modules": false
-          }
-        ]
-      ]
-    },
-    "es": {
-      "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "loose": true,
-            "modules": false
-          }
-        ]
-      ]
-    },
-    "lib": {
-      "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "loose": true
-          }
-        ]
-      ]
-    },
-    "production": {
-      "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "loose": true,
-            "modules": false
-          }
-        ]
-      ]
-    },
-    "test": {
-      "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "loose": true
-          }
-        ]
-      ]
-    }
-  }
-}
