--- .eslintrc
@@ -1,16 +1,0 @@
-{
-  "extends": "rapid7/browser",
-  "globals": {
-    "__dirname": true,
-    "global": true,
-    "module": true,
-    "process": true,
-    "require": true
-  },
-  "parser": "babel-eslint",
-  "parserOptions": {
-    "ecmaVersion": 2017,
-    "sourceType": "module"
-  },
-  "rules":{}
-}
