--- .npmignore
@@ -1,13 +1,0 @@
-.babelrc
-.eslintrc
-.git
-.gitignore
-.idea
-.npmignore
-.nyc_output
-coverage
-DEV_ONLY
-node_modules
-src
-test
-webpack
