--- .npmrc
@@ -1,1 +1,0 @@
-scripts-prepend-node-path=true
