+++ package.json
@@ -66,20 +66,15 @@
   "scripts": {
     "build": "npm run clean && npm run build:dist && npm run build:types",
     "build:dist": "NODE_ENV=production rollup -c config/rollup.config.js",
-    "build:minified": "NODE_ENV=production webpack --progress --colors --config=webpack/webpack.config.minified.js",
     "build:types": "pti fix-types -l dist",
     "clean": "rm -rf dist",
     "clean:cjs": "rm -rf dist/cjs",
-    "clean:dist": "rimraf dist",
     "clean:es": "rm -rf dist/es",
-    "clean:lib": "rimraf lib",
     "clean:umd": "rm -rf dist/umd",
     "dev": "vite --config=config/vite.config.ts",
-    "dist": "npm run clean:dist && npm run build && npm run build:minified",
     "format": "prettier . --log-level=warn --write",
     "format:check": "prettier . --log-level=warn --check",
     "lint": "eslint --max-warnings=0",
-    "lint:fix": "NODE_ENV=test eslint src --fix",
     "release:alpha": "release-it --config=config/release-it/alpha.json",
     "release:beta": "release-it --config=config/release-it/beta.json",
     "release:rc": "release-it --config=config/release-it/rc.json",
