--- webpack/webpack.config.js
@@ -1,46 +1,0 @@
-'use strict';
-
-const path = require('path');
-const webpack = require('webpack');
-
-const ROOT = path.resolve(__dirname, '..');
-
-module.exports = {
-  devtool: '#source-map',
-
-  entry: path.join(ROOT, 'src/index.js'),
-
-  mode: 'development',
-
-  module: {
-    rules: [
-      {
-        enforce: 'pre',
-        include: [path.resolve(ROOT, 'src')],
-        loader: 'eslint-loader',
-        options: {
-          emitError: true,
-          failOnError: true,
-          failOnWarning: true,
-          formatter: require('eslint-friendly-formatter'),
-        },
-        test: /\.js$/,
-      },
-      {
-        include: [path.resolve(ROOT, 'src'), /DEV_ONLY/],
-        loader: 'babel-loader',
-        test: /\.js$/,
-      },
-    ],
-  },
-
-  output: {
-    filename: 'pathington.js',
-    library: 'pathington',
-    libraryTarget: 'umd',
-    path: path.resolve(ROOT, 'dist'),
-    umdNamedDefine: true,
-  },
-
-  plugins: [new webpack.EnvironmentPlugin(['NODE_ENV'])],
-};
