--- webpack/webpack.config.minified.js
@@ -1,22 +1,0 @@
-'use strict';
-
-const webpack = require('webpack');
-
-const defaultConfig = require('./webpack.config');
-
-module.exports = Object.assign({}, defaultConfig, {
-  devtool: undefined,
-
-  mode: 'production',
-
-  output: Object.assign({}, defaultConfig.output, {
-    filename: 'pathington.min.js',
-  }),
-
-  plugins: defaultConfig.plugins.concat([
-    new webpack.LoaderOptionsPlugin({
-      debug: false,
-      minimize: true,
-    }),
-  ]),
-});
